requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
croniter==1.3.5
Django==4.0.5
django-crispy-forms==1.14.0
django-cron==0.6.0
django-crontab==0.7.1
document==1.0
mysqlclient==2.1.0
null==0.6.1
password==0.2
python-dateutil==2.8.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
